modules/core/registry-api/src/main/java/com/liferay/registry/collections/ServiceTrackerCustomizers.java
/**
 * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
 * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
 */

package com.liferay.registry.collections;

import com.liferay.registry.Registry;
import com.liferay.registry.RegistryUtil;
import com.liferay.registry.ServiceReference;
import com.liferay.registry.ServiceTrackerCustomizer;

import java.util.Map;

/**
 * @author Carlos Sierra Andrés
 */
public class ServiceTrackerCustomizers {

	public static <S> ServiceTrackerCustomizer<S, ServiceWrapper<S>>
		serviceWrapper() {

		return new ServiceTrackerCustomizer<S, ServiceWrapper<S>>() {

			@Override
			public ServiceWrapper<S> addingService(
				final ServiceReference<S> serviceReference) {

				Registry registry = RegistryUtil.getRegistry();

				final S service = registry.getService(serviceReference);

				if (service == null) {
					return null;
				}

				try {
					final Map<String, Object> properties =
						serviceReference.getProperties();

					return new ServiceWrapper<S>() {

						@Override
						public Map<String, Object> getProperties() {
							return properties;
						}

						@Override
						public S getService() {
							return service;
						}

					};
				}
				catch (Throwable throwable) {
					registry.ungetService(serviceReference);

					throw throwable;
				}
			}

			@Override
			public void modifiedService(
				ServiceReference<S> serviceReference,
				ServiceWrapper<S> serviceWrapper) {
			}

			@Override
			public void removedService(
				ServiceReference<S> serviceReference,
				ServiceWrapper<S> serviceWrapper) {

				Registry registry = RegistryUtil.getRegistry();

				registry.ungetService(serviceReference);
			}

		};
	}

	public interface ServiceWrapper<S> {

		public Map<String, Object> getProperties();

		public S getService();

	}

}